backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
replicate>=1.0.0
httpx>=0.27.0
pydub>=0.25.1
librosa>=0.10.2
# demucs and basic-pitch both pull in torch — install separately if conflicts arise:
#   pip install torch torchaudio --index-url https://download.pytorch.org/whl/cpu
#   pip install demucs
#   pip install basic-pitch
demucs>=4.0.1
basic-pitch>=0.3.3
